Move the IMove interface to Mods.Common.
